Standardize money on integer cents across Supabase and Dart (issue #8) (#185)
Converts every decimal-dollar money column (menu_items.price,
grocery_products/pharmacy_products.unit_price, and every
subtotal/delivery_fee/tax/total/unit_price column on the food/grocery/
pharmacy order tables) to integer cents, rewrites place_food_order/
place_grocery_order/place_pharmacy_order to compute entirely in cents,
and updates every Dart model, controller, and screen that reads or
writes money to match -- parsing/storing as int cents and converting to
dollars only at display time via a new AppMoney.formatCents helper.

wallet_transactions.amount was already integer cents and is untouched;
the legacy orders/cart_items tables this issue also asks about were
already dropped as dead code by issue #75, before this change.

+++ lib/services/pharmacy/models/pharmacy_cart_item.dart
@@ -6,7 +6,8 @@ class PharmacyCartItem {
   final PharmacyProduct product;
   final int quantity;
 
-  double get total => product.unitPrice * quantity;
+  /// In integer cents — see issue #8.
+  int get total => product.unitPrice * quantity;
 
   PharmacyCartItem copyWith({int? quantity}) {
     return PharmacyCartItem(
